# apps/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.CreateStoreRequest:
    properties:
      address:
        maxLength: 255
        type: string
      category_ids:
        items:
          type: integer
        type: array
      city:
        maxLength: 100
        type: string
      country:
        maxLength: 50
        type: string
      cover_image_url:
        maxLength: 255
        type: string
      description:
        maxLength: 2000
        type: string
      hours:
        items:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.StoreHourRequest'
        type: array
      images:
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      menu_images:
        items:
          type: string
        type: array
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 50
        type: string
      state:
        maxLength: 100
        type: string
      website:
        maxLength: 255
        type: string
      zip_code:
        maxLength: 20
        type: string
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.StoreHourRequest:
    properties:
      close_time:
        maxLength: 10
        type: string
      day_of_week:
        type: integer
      is_closed:
        type: boolean
      open_time:
        maxLength: 10
        type: string
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.UpdateStoreRequest:
    properties:
      address:
        maxLength: 255
        type: string
      category_ids:
        items:
          type: integer
        type: array
      city:
        maxLength: 100
        type: string
      country:
        maxLength: 50
        type: string
      cover_image_url:
        maxLength: 255
        type: string
      description:
        maxLength: 2000
        type: string
      hours:
        items:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.StoreHourRequest'
        type: array
      images:
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      menu_images:
        items:
          type: string
        type: array
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 50
        type: string
      state:
        maxLength: 100
        type: string
      website:
        maxLength: 255
        type: string
      zip_code:
        maxLength: 20
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for RevieU
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RevieU Core API
  version: "1.0"
paths:
...
  /merchant/stores:
    get:
      description: Returns stores owned by the authenticated merchant user
      parameters:
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List current merchant stores
      tags:
      - store
    post:
      consumes:
      - application/json
      description: Creates a new store for the authenticated merchant
      parameters:
      - description: Create store request
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.CreateStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a store
      tags:
      - store
  /merchant/stores/{id}:
    delete:
      description: Soft-deletes a merchant-owned store and its bound coupons
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a store
      tags:
      - store
    patch:
      consumes:
      - application/json
      description: Updates a store for the authenticated merchant
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update store request
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_store_dto.UpdateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a store
      tags:
      - store
  /merchant/stores/{id}/activate:
    post:
      description: Marks a merchant-owned store as published
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Activate a store
      tags:
      - store
...
  /merchant/stores/{id}/deactivate:
    post:
      description: Marks a merchant-owned store as hidden
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a store
      tags:
      - store
...
  /stores:
    get:
      description: Returns a list of stores
      parameters:
      - description: Category name or ID
        in: query
        name: category
        type: string
      - description: Latitude
        in: query
        name: lat
        type: number
      - description: Longitude
        in: query
        name: lng
        type: number
      - description: Minimum average rating
        in: query
        name: rating
        type: number
      - description: Search radius in KM (default 20)
        in: query
        name: radius_km
        type: number
      - description: Cursor for pagination (store id)
        in: query
        name: cursor
        type: integer
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List stores
      tags:
      - store
  /stores/{id}:
    get:
      description: Returns a store by ID
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get store detail
      tags:
      - store
...
  /stores/{id}/hours:
    get:
      description: Returns operating hours for a store
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get store hours
      tags:
      - store
  /stores/{id}/reviews:
    get:
      description: Returns reviews for a store
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cursor for pagination (review id)
        in: query
        name: cursor
        type: integer
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get store reviews
      tags:
      - store
...
schemes:
- http
- https
swagger: "2.0"